1st implementation finish. animations.js
I finished a crude implementation of the raindrop feature. 
Not exactly great.
Workarounds for what I wanted to do were few. I tried a few, but most were outdated, and I wanted this to be a component that can be easily updated in the future, in case I use it. 
Workarounds included jQuery.ripples, using canvas and some complicated JS algorithms, and using an art program to create an actual animation and then running the animation.